an other integer to float

+++ iridescentBasicSynth.ino
@@ -48,8 +48,8 @@ iridescentBasicSynth::iridescentBasicSynth(AudioMixer4 *l, AudioMixer4 *r, Bounc
   ///////////
   
   for (int i = 0; i < TOTAL_VOICES; ++i) {
-    waveformMod[i].mod.phaseModulation((float) 0.33 * rad * i);
-    waveformMod[i].mod.amplitude(1);
+    waveformMod[i].mod.phaseModulation((float) 0.33 * rad * (float) i);
+    waveformMod[i].mod.amplitude(1.0);
     waveformMod[i].mod.begin(waveform_type);
     waveformMod[i].mod.arbitraryWaveform(SawDerange, 12543.0);
     //waveformMod[i].mod.frequencyModulation(3.5);
